requirements.txt
anyio==4.0.0
bardapi==0.1.33
beautifulsoup4==4.12.2
browser-cookie3==0.19.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
deep-translator==1.11.4
exceptiongroup==1.1.3
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
jeepney==0.8.0
langdetect==1.0.9
lz4==4.3.2
numpy==1.25.2
pandas==2.1.0
proto-plus==1.22.3
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tzdata==2023.3
urllib3==1.26.16
